Use non-0 exit code when quitting
Currently hunk uses `:qa` to exit when the quit keybinding is pressed.
This works fine during a split because the output directory will be
identical to the current changeset but this does not work when
performing an operation like squash.

This is because the squash operation will proceed with the entire
changeset instead of being cancelled.

This updates the quit binding to use `:cq` instead, which exits with a
status code of 1.