Manually construct diff buffers
This change replaces the current behaviour of calling `:edit <file>`
with manually constructed buffers.

This gives hunk more control over the buffer names and properties.

The main issue this resolves is the buffer name inheriting the name of
the file from the temporary diff paths. This doesn't play nicely with
lsp clients and often causes lots of diagnostics noise in the file.

By manually defining the buffer names we can ensure lsp clients start in
the correct root directory.