Pass reference to `tree` when calling action fns
When the tree first loads it attempts to call the `on_preview` hook.
This hook currently references the `tree` var which is only bound after
the tree has completed loading.

This introduces a race and results in the `tree` var being nil for
buffers that are opened on init.

This is fixed in this change by passing an explicit reference to itself
when the tree component called `on_preview`.

Fixes #13